Crate liblinkspace

Linkspace - a general purpose supernet

Supernet [ˈsü-pərˌnet] A self-referential multi-participant data organization protocol whose primary addressing method uses hashes instead of endpoint identifiers. A communication protocol where the method of exchange is an extraneous concern. e.g. git, bitcoin, nostr, linkspace

In a supernet anybody can talk about data, instead of talking at a server.

A multi-copy data-first systems. A supernet is ideal when multiple participants want to own and administrate a (part of a) digital system. This is oppose to the current technologies that have users contact a single host, which acts as the de facto administrator by virtue of hosting the data.

Linkspace is supernet with the following highlights:

Small and powerful API

Fast (Blake3, no JSON/Base64 encoding, well aligned fields)

Path (URL like) addressable packets.

Group/Domain split

To write applications for a supernet requires a different perspective compared to managing sockets. The challenge is defining a super structure that can work across time and space with noisy incomplete data. Or treat it as a shared folder. That works as well.

In return a supernet provides a lot of useful properties including: Serverless, free, extensible, reusable, adaptable, scalable, lockin-resistant, lockout-resistant, verifyable, optional accountability, inherent privacy,fault tolerant etc.

Checkout Basics for a simple introduction. Download to give it a try and say hi on the test group. Checkout the Guide if you’re up for some programming.

The packet format and index is stable, but expect some unimplemented feature and rough edges.

The functions re-exported below is essentially the entire liblinkspace interface. Bindings in other languages follow the same pattern.

prelude includes some additional utilities. Some internals structs defs are currently leaking and will be removed.

Re-exports

pub use point::lk_datapoint;

pub use point::lk_keypoint;

pub use point::lk_linkpoint;

pub use abe::lk_encode;

pub use abe::lk_eval;

pub use abe::lk_split_abe;

pub use query::lk_query;

pub use query::lk_query_parse;

pub use query::lk_query_print;

pub use query::lk_query_push;

pub use query::Query;

pub use key::lk_key;

pub use linkspace::lk_get;

pub use linkspace::lk_open;

pub use linkspace::lk_process;

pub use linkspace::lk_process_while;

pub use linkspace::lk_save;

pub use linkspace::lk_stop;

pub use linkspace::lk_watch;

pub use linkspace::Linkspace;

pub use conventions::lk_pull;

pub use misc::cb;

Modules

abe

ascii byte expression utilities

consts

conventions

A set of functions that adhere to conventions

key

linkspace

misc

point

prelude

query

varctx

Functions with a custom eval context

Structs

SigningKey

Constants

PRIVATE

PUBLIC

Traits

PktHandler

Callbacks stored in a Linkspace instance. use misc::cb to impl from function

Type Definitions

LkError

LkResult