Module liblinkspace::abe

ascii byte expression utilities

ABE is a byte templating language. See guide#ABE to understand its use and indepth explanation. *

Modules

ctx

Custom context for use in varctx

Statics

DEFAULT_PKT

default fmt in many cases and output for [pkt]

Functions

lk_encode

encode bytes as an abe that evaluate back to bytes.

lk_eval

Evaluate an expression and return the bytes

lk_split_abe

Exec callback for each expr between control characters (‘:’, ‘/’, ‘\n’, ‘\t’). The last delimiter can be ‘\0’.

lk_try_encode

lk_encode with Err on: