Function liblinkspace::linkspace::lk_watch

pub fn lk_watch(
    lk: &Linkspace,
    query: &Query,
    cb: impl PktHandler + 'static
) -> LkResult<u32>

watch packets matching the query - both already in the db and new packets on arrival

Calls cb for each matching packet. If the query contains the watch option ( e.g. ‘:watch:example’ ) the cb is also called for all new packets during [lk_process] and [lk_process_while]. The watch is dropped when

the cb returns ‘break’ ( usually false )

[lk_stop] is called with the matching id

the predicate set will never match again ( ‘i’ counters and recv )

returns the number matches in the local index. i.e. the number of times cb was called immediatly. *