Trait liblinkspace::prelude::NetPkt

pub trait NetPkt: Debug {
    // Required methods
    fn as_point(&self) -> &dyn Point;
    fn hash_ref(&self) -> &B64<[u8; 32]>;
    fn net_header_ref(&self) -> &NetPktHeader;
    fn recv(&self) -> Option<U64>;
    fn byte_segments(&self) -> ByteSegments<'_>;

    // Provided methods
    fn net_header_mut(&mut self) -> Option<&mut NetPktHeader> { ... }
    fn as_netbox(&self) -> Box<NetPktFatPtr, Global> { ... }
    fn as_netarc(&self) -> NetPktArc  { ... }
}

A trait to access fields of a net pkt. Auto impls NetPktExt, Point, and PointExt.

A NetPkt is the combination of Point, LkHash, NetPktHeader. the trait is impl for various layouts such as NetPktBox, NetPktArc.

There are wrapping structs in as [reroute] that provide extended options.

Required Methods

fn as_point(&self) -> &dyn Point

fn hash_ref(&self) -> &B64<[u8; 32]>

fn net_header_ref(&self) -> &NetPktHeader

fn recv(&self) -> Option<U64>

fn byte_segments(&self) -> ByteSegments<'_>

Provided Methods

fn net_header_mut(&mut self) -> Option<&mut NetPktHeader>

fn as_netbox(&self) -> Box<NetPktFatPtr, Global>

fn as_netarc(&self) -> NetPktArc